Move tests for `IntervalIdentifier` into separate test module.
